feat: option to disable queries on server
