+++ src/composables/useConvexClient.ts
@@ -1,31 +1,14 @@
-import type { ConvexClient, ConvexHttpClient } from 'convex/browser'
-import type { ConvexVueContext } from '../plugin'
-import { inject } from 'vue'
+import type { ConvexClient } from 'convex/browser'
+import { useConvexContext } from './useConvexContext'
 
 /**
  * Returns the Convex client instance.
  */
 export function useConvexClient(): ConvexClient {
-  const convexVueContext = inject<ConvexVueContext>('convex-vue')
-  if (!convexVueContext)
-    throw new Error('Context not found')
+  const convexVueContext = useConvexContext()
 
   if (!convexVueContext.clientRef.value)
     throw new Error('Client not initialized')
 
   return convexVueContext.clientRef.value
-}
-
-/**
- * Returns the Convex HTTP client instance.
- */
-export function useConvexHttpClient(): ConvexHttpClient {
-  const convexVueContext = inject<ConvexVueContext>('convex-vue')
-  if (!convexVueContext)
-    throw new Error('Context not found')
-
-  if (!convexVueContext.httpClientRef.value)
-    throw new Error('HTTP client not initialized')
-
-  return convexVueContext.httpClientRef.value
 }
